pyproject.toml
[build-system]
requires = [ "setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zebrafy"
version = "0.1.0"
description = "Python library for converting Zebra Programming Language (ZPL) to and from PDF, HTML, and images"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Manufacturing",
    "Topic :: Printing",
]
dependencies = ["pillow", "pypdfium2<=4.18.0"]
requires-python = ">=3.7"

[project.optional-dependencies]
test = [
    "pytest-cov",
]

docs = [
    "sphinx",
    "sphinx_autodoc_typehints",
    "sphinx_rtd_theme>=1.2.0",
]

[[project.authors]]
name = "Miika Nissi"
email = "miika@miikanissi.com"

[[project.maintainers]]
name = "Miika Nissi"
email = "miika@miikanissi.com"

[project.license]
text = "LGPLv3"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/miikanissi/zebrafy"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = [ "zebrafy"]
namespaces = false

[tool.pytest.ini_options]
addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 95"

[tool.coverage.run]
source = ["zebrafy"]
